LayoutTests/platform/mac/imported/w3c/web-platform-tests/svg/shapes/scripted/disabled-shapes-not-hit-expected.txt

CONSOLE MESSAGE: line 70: Error: Invalid negative value for <rect> attribute width="-10"
CONSOLE MESSAGE: line 70: Error: Invalid negative value for <rect> attribute height="-10"
CONSOLE MESSAGE: line 70: Error: Invalid negative value for <circle> attribute r="-10"
CONSOLE MESSAGE: line 96: TypeError: null is not an object (evaluating 'document.body.appendChild')
#PID UNRESPONSIVE - WebKitTestRunner (pid 25146)
FAIL: Timed out waiting for notifyDone to be called

#EOF
#EOF
